Matsuura: Update Kernel Script Builder
During AnyKernel2 can't be ported to flamingo cause of problem when extracting
ramdisk, so kernel now will builded as anykernel1.

Kernel script is getting update to make 2 option to build kernel until anykernel2
is getting known to status on development. For now is only suspended not abandoned
so it should to keep it available but not to compile it.

Signed-off-by: Dicky Herlambang <herlambangdicky5@gmail.com>